Add instances for SNat, SSymbol, and SChar
These were introduced to `GHC.TypeNats` and `GHC.TypeLits` in `base-4.18.0.0`.
They did not receive `Eq` or `Ord` instances until `base-4.19.0.0`, however,
which are required for the superclasses of `EqP` and `OrdP`. To account for
this, I introduce a conditional dependency on `base-orphans`, which backports
the `Eq` and `Ord` instances to older versions of `base`.